* start getting the layout rendering properly and add the .bg-body class to <body> tag
