+++ .storybook/main.js
@@ -16,23 +16,8 @@ module.exports = {
   ],
   webpackFinal: async (config) => {
     config.resolve.alias['lib'] = path.resolve(__dirname, '../lib')
-    // config.module.rules.push({
-    //   test: /\.css$/,
-    //   use: [
-    //     {
-    //       loader: 'postcss-loader',
-    //       options: {
-    //         ident: 'postcss',
-    //         plugins: [
-    //           require('postcss-easy-import'),
-    //           require('tailwindcss'),
-    //           require('autoprefixer')
-    //         ]
-    //       }
-    //     }
-    //   ],
-    //   include: path.resolve(__dirname, '../')
-    // })
+    config.resolve.alias['stories'] = path.resolve(__dirname, '../stories')
+
     return config
   }
 }
