+++ postcss.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   plugins: {
+    'postcss-easy-import': {},
     tailwindcss: {},
-    autoprefixer: {},
+    'autoprefixer': {}
   },
 }
