--- postcss.config.js.bak
@@ -1,7 +1,0 @@
-module.exports = {
-  plugins: {
-    'postcss-easy-import': {},
-    tailwindcss: {},
-    'autoprefixer': {}
-  },
-}
