+++ stories/assets/globals.css
@@ -1,3 +1,5 @@
+@import '../../lib/styles/index.css';
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
