--- tailwind.config.js.bak
@@ -1,10 +1,0 @@
-const pooltogetherTheme = require('./config')
-
-module.exports = pooltogetherTheme({
-  purge: [],
-  theme: {
-    extend: {},
-  },
-  variants: {},
-  plugins: [],
-})
